Use mutex instead of binary semaphore.
Remove debugaids.o from link.

Fix format string for pointer type.

Make constant 1's unsigned, since left shift of 1 by 31 places cannot be represented in type int.

+++ sw_projects/P2_app/OutHighPriority.c
@@ -50,7 +50,7 @@ void *OutgoingHighPriority(void *arg)
   struct ThreadSocketData *ThreadData;            // socket etc data for this thread
   struct sockaddr_in DestAddr;                    // destination address for outgoing data
   bool InitError = false;
-  int Error;
+  ssize_t Error;
   uint8_t Byte;                                   // data being encoded
   uint16_t Word;                                  // data being encoded
   bool ATUTuneRequest = false;
